Modularize 'paginationConnection' in utils/pageConnection.ts
