+++ src/types/models/User.ts
@@ -3,61 +3,7 @@ import { intArg, objectType, stringArg } from '@nexus/schema';
 import { Post } from './Post';
 import { createPageEdges } from '../../utils/pageEdge';
 import { getUserId } from '../../utils';
-
-const cursorConnection = objectType({
-  name: 'cursorConnection',
-  definition(t) {
-    t.string('cursor');
-    t.int('page');
-    t.boolean('isCurrent');
-  },
-});
-
-const pageCursorsConnection = objectType({
-  name: 'pageCursorsConnection',
-  definition(t) {
-    t.field('previous', {
-      type: cursorConnection,
-      nullable: true,
-    });
-    t.field('first', {
-      type: cursorConnection,
-      nullable: true,
-    });
-    t.field('last', {
-      type: cursorConnection,
-      nullable: true,
-    });
-    t.list.field('around', {
-      type: cursorConnection,
-      nullable: true,
-    });
-  },
-});
-
-const pageEdgesConnection = objectType({
-  name: 'pageEdgesConnection',
-  definition(t) {
-    t.string('cursor');
-    t.field('node', {
-      type: Post,
-    });
-  },
-});
-
-const paginationConnection = objectType({
-  name: 'paginationConnection',
-  definition(t) {
-    t.list.field('pageEdges', {
-      type: pageEdgesConnection,
-      nullable: true,
-    });
-    t.field('pageCursors', {
-      type: pageCursorsConnection,
-      nullable: true,
-    });
-  },
-});
+import { paginationConnection } from '../../utils/pageConnection';
 
 export const Profile = objectType({
   name: 'Profile',
